limit delete item according to permissions

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -52,8 +52,20 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // 1. find the item
-    const item = await ctx.db.query.item({ where }, `{id title}`);
+    const item = await ctx.db.query.item({ where }, `{id title user { id }}`);
     // 2. Check if they own the item or have the permissions
+    const ownsItem = item.user.id === ctx.request.userId;
+
+    // ['DOG', 'CAT'].some(pet => ['BIRD', 'SNAKE'].includes(pet)); // false
+    // ['DOG', 'CAT'].some(pet => ['BIRD', 'SNAKE', 'CAT'].includes(pet)); // true
+    // At least one of the permissions is in the required ones!
+    const hasPermissions = ctx.request.user.permissions.some((permission) =>
+      ["ADMIN", "ITEMDELETE"].includes(permission)
+    );
+
+    if (!ownsItem && !hasPermissions) {
+      throw new Error(`You don't have permission to do that!`);
+    }
     // 3. Delete it!
     return ctx.db.mutation.deleteItem({ where }, info);
   },
